Allow the ability to completely disable the marketplace.
