+++ src/components/Header.astro
@@ -4,7 +4,7 @@ import { getLangFromUrl, useTranslations } from "../i18n/utils";
 import { isMobileNavOpen } from "../store.js";
 import HeaderButton from "./HeaderButton.astro";
 import Logo from "./Logo.astro";
-
+import { MARKETPLACE_ENABLED } from "astro:env/client";
 const lang = getLangFromUrl(Astro.url);
 const t = useTranslations(lang);
 ---
@@ -40,12 +40,14 @@ const t = useTranslations(lang);
           class="h-6 w-6 text-text-color transition duration-500 group-hover:text-text-hover-color md:h-6 md:w-6"
         />
       </HeaderButton>
-      <HeaderButton text={t("header.catalog")} route={`/catalog/1`}>
-        <Icon
-          name="ph:shopping-bag-open-fill"
-          class="h-6 w-6 text-text-color transition duration-500 group-hover:text-text-hover-color md:h-6 md:w-6"
-        />
-      </HeaderButton>
+      {MARKETPLACE_ENABLED &&
+        <HeaderButton text={t("header.catalog")} route={`/catalog/1`}>
+            <Icon
+                name="ph:shopping-bag-open-fill"
+                class="h-6 w-6 text-text-color transition duration-500 group-hover:text-text-hover-color md:h-6 md:w-6"
+             />
+        </HeaderButton>
+      }
       <HeaderButton text={t("header.morelinks")}>
         <Icon
           name="ph:link-bold"
